Replaced old GUI code.
